requirements.txt
baidu-aip
pydub
srt
Flask